add finalization span and a variety of other fixes
